Add files via upload
Added Visor.py a simple library that can be used to control a protogens visor built from neopixels. It has a dev mode so that it can be tested and developed with outside a raspberry pi. it contains 3 dictionary to choose the shape of the mouth, eyes, nose, and the colors can be set